Add struct that stores task info needed to restart. Change where oom_restart() called
The needed data for restarting a task is now saved within oom_kill_process()
rather than after it returns. Waiting until after it returns leads to a race
condition where oom_restart() would try to read the killed tasks task_struct
after the kernel kad killed the task. The task_struct was not gauranteed to
exist when oom_restart() tried to access it.