Tokenize cmdline required for kernel_execve. Change location of oom_restart call.
The cmdline passed to kernel_execve must be stored as argv[] is in C programs.
The location of oom_restart call now allows process to be killed and memory
reclaimed before attempting to determine if it should be restarted. Calling it
before the memory had been reclaimed led to the system hanging requiring a
reboot.