Find source of hangs. Minor changed to slow_java.sh.
Using kgdb and refering to logs, it didn't make sense that the kernel was
hanging. The VM was set to boot in console mode and it was found the X server
was the source of the hangs. Running in console mode, both the vanilla kernel
and the modified kernel can run for hours being repeatedly put in an OOM
situation. The modified kernel will start logger most of the time with a sizable
minority of calls to oom_restart() resulting in an "out of memory" error.