Add a second implementation of netlink lkm.
The module and userspace task seem to see eachother as the restarter task cannot
be started until the lkm is loaded or a "protocol not supported" error is
returned from the restarter task. However, when an oom situation is created, the
lkm's attempt to send a multicast message to MYGRP fails with error "no such
process".